fix(install-test-server): make compatible to Ubuntu 24.04
This needs the following further manual patch to work correctly:

```
diff --git a/roles/eaas-docker/tasks/main.yaml b/roles/eaas-docker/tasks/main.yaml
index 801f9ae..2b54b95 100644
--- a/roles/eaas-docker/tasks/main.yaml
+++ b/roles/eaas-docker/tasks/main.yaml
@@ -10,13 +10,3 @@
 - name: install docker-compose
   when: "unsupported_os is undefined"
   import_tasks: docker-compose.yaml
-
-- name: install python bindings
-  become: yes
-  when: "unsupported_os is undefined"
-  pip:
-    name:
-    - requests<2.29
-    - urllib3<2
-    - docker
-    state: present
```

+++ scripts/install-dependencies.sh
@@ -10,6 +10,13 @@ apt-get update && DEBIAN_FRONTEND=noninteractive apt-get install -y \
 DEBIAN_FRONTEND=noninteractive apt-get install -y python-is-python3 \
   || update-alternatives --install /usr/bin/python python /usr/bin/python3 100
 
+. /etc/os-release
+
+if [ "${NAME-}" = "Ubuntu" ] && [ "${VERSION_ID-}" = "24.04" ]; then
+  DEBIAN_FRONTEND=noninteractive apt-get install -y docker-compose ansible-core
+  exit
+fi
+
 # HACK: make compatible to Python 3.11+
 rm -f /usr/lib/python3.*/EXTERNALLY-MANAGED
 pip3 install --break-system-packages --no-build-isolation docker-compose "docker<7" ansible
